pyproject.toml
[project]
name = "webhook-handler"
version = "0.1.0"
description = "Welkin Webhook Handler"
readme = "README.md"
requires-python = "==3.12.2"
package-mode = false
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "pydantic-settings>=2.9.1",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "black==25.1.0",
]

[tool.black]
line-length = 120
exclude = '''
/(
  | .venv
  | data
  | logger
)/
'''
